Move tslint to tasks/lint
