+++ bokehjs/gulp/tasks/index.ts
@@ -5,6 +5,7 @@ import "./default"
 import "./defaults"
 import "./examples"
 import "./install"
+import "./lint"
 import "./scripts"
 import "./styles"
 import "./test"
